Add support for extensible scheduler compilation and provide sample recipe
This change adds the neccessary kernel config changes required for running an extensible scheduler. It also provides 2 attempts at recipes to build an extensible scheduler. However, the current state of cross-compiler compatibility for both build systems seems very bad. The meson based userspace build system supports inherent cross compilation through the meson.cross file, but the build process uses clang as a binary (not as CC) for bpf compilation. Thus, cross compilation flags are not passed down properly. The build system (autotools) provided in the kernel source seemed to be better (although it was recommended to be used only for testing purposes). However, it uses make without wrapping $CC in quotes. This is again an issue since $CC contains many flags during cross-compilation. Thus, both the recipes are not yet production ready but can be used as a reference to come up with a working recipe.