Update to most recent upstream
This change partially changes the upstream to the whinlatter branch as the meta-ti layer is compatile with the same in its most recent release. In the future, care will be taken to pinpoint exact commits instead of branch names in the kas config, so that such breaking changes are not very frequent.